feat(application): migrate workspace-init-remote to shared application API
Move remote workspace initialization from domain workflow into the
application use-case layer, following the established pattern from
other migrated use cases (move_item, create_item, etc.).

The CLI adapter now imports from application/use_cases instead of
domain/workflows. Local-mode workspace init is unaffected.
